+++ Code/Mantid/Framework/DataHandling/test/ManagedRawFileWorkspace2DTest.h
@@ -187,9 +187,9 @@ public:
     // Test one to many mapping, for example 10 pixels contribute to spectra 2084
     TS_ASSERT_EQUALS(map.ndet(2084),10);
     // Check the id number of all pixels contributing
-    std::vector<int> detectorgroup;
+    std::vector<int64_t> detectorgroup;
     detectorgroup=map.getDetectors(2084);
-    std::vector<int>::const_iterator it;
+    std::vector<int64_t>::const_iterator it;
     int pixnum=101191;
     for (it=detectorgroup.begin();it!=detectorgroup.end();it++)
     TS_ASSERT_EQUALS(*it,pixnum++);
@@ -198,7 +198,7 @@ public:
     // Test that number of pixel=0
     TS_ASSERT_EQUALS(map.ndet(5),0);
     // Test that trying to get the Detector throws.
-    std::vector<int> test = map.getDetectors(5);
+    std::vector<int64_t> test = map.getDetectors(5);
     TS_ASSERT(test.empty());
     
     AnalysisDataService::Instance().remove(outputSpace);
