+++ Code/Mantid/Framework/DataObjects/test/SpecialWorkspace2DTest.h
@@ -44,7 +44,7 @@ public:
     TS_ASSERT_EQUALS( ws->blocksize(), 1);
     TS_ASSERT_EQUALS( ws->getInstrument()->getName(), "basic"); // Name of the test instrument
     TS_ASSERT_EQUALS( ws->spectraMap().nElements(), 45);
-    std::vector<int> dets = ws->spectraMap().getDetectors(0);
+    std::vector<int64_t> dets = ws->spectraMap().getDetectors(0);
     TS_ASSERT_EQUALS(dets.size(), 1);
 
     TS_ASSERT_EQUALS( ws->getDetectorID(0), 1);
